Seaborn and parquet
Turns out that the super-fancy-good-looking plotting library "ploty" keeps ALL the data points for a plot, and reloads them on page load etc. It also uses json... to pass the data from python to the javascript library.
This is how it makes nice interactive plots where i can drag things around.
But this means that for instance the histogram of the nodata-radiuses for the whole 3dbag is like a 400mb object which is reloaded each time i sneeze